Check formulas for computing count
