+++ internal/server/count/count.go
@@ -19,6 +19,7 @@ import (
 
 	pb "github.com/datacommonsorg/mixer/internal/proto"
 	"github.com/datacommonsorg/mixer/internal/server/cache"
+	"github.com/datacommonsorg/mixer/internal/server/statvar/formula"
 	"github.com/datacommonsorg/mixer/internal/sqldb/sqlquery"
 	"github.com/datacommonsorg/mixer/internal/store"
 	"github.com/datacommonsorg/mixer/internal/store/bigtable"
@@ -26,12 +27,7 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
-// Count checks if entities have data for stat vars and stat var groups.
-//
-// Returns a two level map from sv/svg dcid to entity dcid to the number of
-// sv with data. For a given sv/svg, if an entity has no data, it will
-// not show up in the second level map.
-func Count(
+func countInternal(
 	ctx context.Context,
 	st *store.Store,
 	cachedata *cache.Cache,
@@ -126,6 +122,65 @@ func Count(
 						}
 						result[ancestor][e] += 1
 					}
+				}
+			}
+		}
+	}
+	return result, nil
+}
+
+// Count checks if entities have data for stat vars and stat var groups.
+//
+// Returns a two level map from sv/svg dcid to entity dcid to the number of
+// sv with data. For a given sv/svg, if an entity has no data, it will
+// not show up in the second level map.
+func Count(
+	ctx context.Context,
+	st *store.Store,
+	cachedata *cache.Cache,
+	svOrSvgs []string,
+	entities []string,
+) (map[string]map[string]int32, error) {
+	result, err := countInternal(ctx, st, cachedata, svOrSvgs, entities)
+	if err != nil {
+		return nil, err
+	}
+	// Check for count for computed observations.
+	// Use counts of formula variables as a heuristic.
+	for _, dcid := range svOrSvgs {
+		formulas, ok := cachedata.SVFormula()[dcid]
+		if !ok {
+			continue
+		}
+		for _, entity := range entities {
+			if _, ok := result[dcid][entity]; ok {
+				continue
+			}
+			for _, f := range formulas {
+				variableFormula, err := formula.NewVariableFormula(f)
+				if err != nil {
+					return nil, err
+				}
+				calculatedCount, err := countInternal(
+					ctx,
+					st,
+					cachedata,
+					variableFormula.StatVars,
+					[]string{entity},
+				)
+				if err != nil {
+					return nil, err
+				}
+				allExist := true
+				for _, entityCount := range calculatedCount {
+					if _, ok := entityCount[entity]; !ok {
+						allExist = false
+						break
+					}
+				}
+				if allExist {
+					result[dcid][entity] = 0
+					break
 				}
 			}
 		}
