+++ internal/server/v2/observation/golden/existence/existence.json
@@ -6,7 +6,8 @@
         "country/CAN": {},
         "geoId/06": {},
         "geoId/0647766": {},
-        "test_entity": {}
+        "test_entity": {},
+        "wikidataId/Q613": {}
       }
     },
     "GenderIncomeInequality_Person_15OrMoreYears_WithIncome": {
@@ -31,6 +32,15 @@
       "by_entity": {
         "geoId/06": {},
         "test_entity": {}
+      }
+    },
+    "Count_Person_Female": {
+      "by_entity": {
+        "country/BRA": {},
+        "country/CAN": {},
+        "geoId/06": {},
+        "geoId/0647766": {},
+        "wikidataId/Q613": {}
       }
     }
   }
